WIP: This would fix missing data from osm.Pbf files
Due to the way Geofabrik handles exports,
https://help.openstreetmap.org/questions/3071/inconsistency-found-in-geofabrik-osm-exports

"What you are seeing is the result of these extracts being created with the Osmosis option --clipIncompleteEntities=true - whenever an object crosses the border of the extract, then only the part of the object that lies inside is actually included in the extract.

Osmosis has two other options how to deal with these situations. One is --clipIncompleteEntities=false which would in your case lead to a way that references all the right nodes, but these nodes would be missing in the file. Geofabrik doesn't do that because the inconsistent files would trip up some users. The other option is --completeWays=true in which Osmosis would collect the extra nodes and add them to the output file even though they lie outside the area in question. This is not used by Geofabrik because it takes too long.

If you cannot work with this file, you will have to use the next larger extract (e.g. france.osm) and then cut out your area of interest yourself with a suitable set of parameters."